create_unified_model.py
#!/usr/bin/env python3
"""
Script to create a unified disease prediction model.
This will consolidate all individual disease models into a single pkl file.
"""

import os
import sys
import numpy as np
import pandas as pd
import joblib
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split

# Add the src/app directory to the Python path
sys.path.append(os.path.join(os.path.dirname(__file__), 'src', 'app'))

# Import our model classes
from model import MultiDiseasePredictor

def load_existing_models():
    """Load all existing individual models and get their feature importances"""
    model_info = {}
    
    # Find all model files
    model_files = [f for f in os.listdir() if f.endswith('_model.pkl') and not f.startswith('unified')]
    scaler_files = [f for f in os.listdir() if f.endswith('_scaler.pkl')]
    
    print(f"Found {len(model_files)} individual model files")
    
    # Load each model and its corresponding scaler
    for model_file in model_files:
        try:
            model_name = model_file.replace('_model.pkl', '')
            scaler_file = f"{model_name}_scaler.pkl"
            
            if scaler_file in scaler_files:
                model = joblib.load(model_file)
                scaler = joblib.load(scaler_file)
                
                # Extract feature importances if available
                feature_importances = None
                if hasattr(model, 'feature_importances_'):
                    feature_importances = model.feature_importances_
                
                model_info[model_name] = {
                    'model': model,
                    'scaler': scaler,
                    'feature_importances': feature_importances
                }
                
                print(f"Loaded model: {model_name}")
        except Exception as e:
            print(f"Error loading model {model_file}: {e}")
    
    return model_info

def create_unified_model(synthetic_data=True):
    """
    Create a unified model from existing individual models or synthetic data.
    
    Args:
        synthetic_data: If True, uses synthetic data to train the model.
                       If False, attempts to use data from existing models.
    """
    # Initialize our disease predictor
    predictor = MultiDiseasePredictor()
    
    # Get all features that will be used in the unified model
    all_features = predictor.all_features
    print(f"Unified model will use {len(all_features)} features: {all_features}")
    
    if not synthetic_data:
        # Try to load existing models to extract information
        model_info = load_existing_models()
        
        if len(model_info) > 0:
            # If we have existing models, use their structure
            # But for simplicity in this implementation, we'll still use synthetic data
            print("Using existing model structure with synthetic training data")
    
    # Generate synthetic data for training
    print("Generating synthetic training data...")
    n_samples = 2000  # Increase sample size for better model performance
    
    # Create more realistic synthetic data distributions
    X = np.zeros((n_samples, len(all_features)))
    feature_distributions = {
        'AGE': (50, 15),  # mean=50, std=15
        'BMI': (25, 5),   # mean=25, std=5
        'Systolic_BP': (120, 15),
        'Diastolic_BP': (80, 10),
        'Heart_Rate': (75, 10),
        'Cholesterol_Total': (200, 40),
        'Blood_Glucose': (100, 25),
        'Oxygen_Saturation': (97, 2),
        'WBC_Count': (7500, 2000),
        'CRP': (1.0, 1.0),
        'Respiratory_Rate': (16, 3),
        'FEV1': (3.0, 0.5),
        'Waist_Circumference': (90, 15),
        'Body_Temperature': (37.0, 0.5),
        'Liver_Enzymes': (30, 10)
    }
    
    # Generate features with appropriate distributions
    X_df = pd.DataFrame(columns=all_features)
    
    for i, feature in enumerate(all_features):
        if feature in feature_distributions:
            mean, std = feature_distributions[feature]
            X_df[feature] = np.random.normal(mean, std, n_samples)
        elif feature == 'GENDER_M':
            X_df[feature] = np.random.choice([0, 1], n_samples)
        elif feature == 'Family_History':
            X_df[feature] = np.random.choice([0, 1], n_samples, p=[0.7, 0.3])
        elif feature == 'Smoking_Status':
            X_df[feature] = np.random.choice([0, 1, 2], n_samples, p=[0.6, 0.2, 0.2])
        else:
            X_df[feature] = np.random.rand(n_samples)
    
    # Generate target variables with dependencies on features
    y_dict = {}
    
    # For each disease, create a more realistic prediction rule based on risk factors
    # Cardiovascular - Hypertension
    bp_risk = (X_df['Systolic_BP'] > 130).astype(int) + (X_df['Diastolic_BP'] > 85).astype(int)
    age_risk = (X_df['AGE'] > 60).astype(int)
    bmi_risk = (X_df['BMI'] > 30).astype(int)
    y_dict['CARDIOVASCULAR_Hypertension'] = ((bp_risk + age_risk + bmi_risk >= 2) | 
                                           (X_df['Systolic_BP'] > 140) | 
                                           (X_df['Diastolic_BP'] > 90)).astype(int)
    
    # Cardiovascular - Heart Attack
    cholesterol_risk = (X_df['Cholesterol_Total'] > 240).astype(int)
    y_dict['CARDIOVASCULAR_Heart_Attack'] = ((age_risk + bp_risk + cholesterol_risk >= 2) | 
                                           (X_df['Systolic_BP'] > 160)).astype(int)
    
    # Metabolic - Diabetes
    glucose_risk = (X_df['Blood_Glucose'] > 126).astype(int) + (X_df['Blood_Glucose'] > 100).astype(int)
    y_dict['METABOLIC_Diabetes'] = ((glucose_risk + bmi_risk + age_risk >= 2) | 
                                  (X_df['Blood_Glucose'] > 126)).astype(int)
    
    # Metabolic - Obesity
    y_dict['METABOLIC_Obesity'] = (X_df['BMI'] > 30).astype(int)
    
    # Respiratory - Asthma
    resp_rate_risk = (X_df['Respiratory_Rate'] > 20).astype(int)
    fev_risk = (X_df['FEV1'] < 2.5).astype(int)
    y_dict['RESPIRATORY_Asthma'] = (resp_rate_risk + fev_risk >= 1).astype(int)
    
    # Respiratory - COPD
    smoking_risk = (X_df['Smoking_Status'] > 0).astype(int)
    y_dict['RESPIRATORY_COPD'] = ((fev_risk + smoking_risk + age_risk >= 2) | 
                                (X_df['FEV1'] < 2.0)).astype(int)
    
    # Oncology - Breast Cancer
    family_risk = (X_df['Family_History'] > 0).astype(int)
    gender_risk = (X_df['GENDER_M'] == 0).astype(int)  # Higher risk for females
    y_dict['ONCOLOGY_Breast_Cancer'] = ((gender_risk + family_risk + age_risk >= 2)).astype(int)
    
    # Oncology - Lung Cancer
    y_dict['ONCOLOGY_Lung_Cancer'] = ((smoking_risk + age_risk + fev_risk >= 2)).astype(int)
    
    # Oncology - Colon Cancer
    y_dict['ONCOLOGY_Colon_Cancer'] = ((family_risk + age_risk + bmi_risk >= 2)).astype(int)
    
    # Infectious - Hepatitis
    liver_risk = (X_df['Liver_Enzymes'] > 40).astype(int)
    y_dict['INFECTIOUS_Hepatitis'] = (liver_risk == 1).astype(int)
    
    # Infectious - Influenza & Flu
    temp_risk = (X_df['Body_Temperature'] > 38.0).astype(int)
    wbc_risk = (X_df['WBC_Count'] > 11000).astype(int)
    y_dict['INFECTIOUS_Influenza'] = ((temp_risk + wbc_risk >= 1)).astype(int)
    y_dict['INFECTIOUS_Flu'] = ((temp_risk + wbc_risk >= 1)).astype(int)
    
    print("Training unified model...")
    
    # Fit the unified scaler
    unified_scaler = StandardScaler()
    X_scaled = unified_scaler.fit_transform(X_df)
    
    # Create disease-specific probability estimators
    unified_model = {
        'scaler': unified_scaler,
        'features': all_features,
        'disease_models': {}
    }
    
    # Train a model for each disease with appropriate train/test split
    for disease_key, y in y_dict.items():
        print(f"Training model for {disease_key}...")
        
        # Split data for training and validation
        X_train, X_test, y_train, y_test = train_test_split(
            X_scaled, y, test_size=0.2, random_state=42)
        
        # Train model
        model = RandomForestClassifier(n_estimators=100, random_state=42)
        model.fit(X_train, y_train)
        
        # Evaluate model
        train_accuracy = model.score(X_train, y_train)
        test_accuracy = model.score(X_test, y_test)
        print(f"  Train accuracy: {train_accuracy:.2f}, Test accuracy: {test_accuracy:.2f}")
        
        # Add to unified model
        unified_model['disease_models'][disease_key] = model
    
    # Save the unified model
    print("Saving unified disease model...")
    joblib.dump(unified_model, 'unified_disease_model.pkl')
    print("Unified model created and saved successfully!")
    
    # Update feature names file if needed
    if not os.path.exists('feature_names.pkl'):
        feature_names = predictor.default_feature_names
        joblib.dump(feature_names, 'feature_names.pkl')
        print("Feature names file created")
    
    return unified_model

if __name__ == "__main__":
    print("Creating unified disease prediction model...")
    create_unified_model()
    print("Done!")