compatible_requirements.txt
numpy>=1.26.0
pandas>=2.0.0
matplotlib>=3.7.1
scikit-learn>=1.2.2
scipy>=1.15.2
seaborn>=0.12.2
streamlit>=1.22.0
altair>=5.5.0
python-dotenv>=1.0.0
mysql-connector-python
imbalanced-learn>=0.13.0
plotly>=6.0.1
